refactor(validation): move validation functions to separate file
