+++ src/services/prompter/application-prompts.js
@@ -1,4 +1,5 @@
 const { default: languages, languageList } = require('../../utils/languages');
+const { validateAppHostname, validatePort } = require('../../utils/validators');
 const AbstractPrompter = require('./abstract-prompter');
 
 class ApplicationPrompts extends AbstractPrompter {
@@ -24,15 +25,7 @@ class ApplicationPrompts extends AbstractPrompter {
           name: 'applicationHost',
           message: "What's the IP/hostname on which your application will be running?",
           default: 'http://localhost',
-          validate: hostname => {
-            if (!/^https?:\/\/.*/i.test(hostname)) {
-              return 'Application hostname must be a valid url.';
-            }
-            if (!/^http((s:\/\/.*)|(s?:\/\/(localhost|127\.0\.0\.1).*))/i.test(hostname)) {
-              return 'HTTPS protocol is mandatory, except for localhost and 127.0.0.1.';
-            }
-            return true;
-          },
+          validate: validateAppHostname,
         });
       }
     }
@@ -47,17 +40,7 @@ class ApplicationPrompts extends AbstractPrompter {
           name: 'applicationPort',
           message: "What's the port on which your application will be running?",
           default: '3310',
-          validate: port => {
-            if (!/^\d+$/.test(port)) {
-              return 'The port must be a number.';
-            }
-
-            const parsedPort = parseInt(port, 10);
-            if (parsedPort > 0 && parsedPort < 65536) {
-              return true;
-            }
-            return 'This is not a valid port.';
-          },
+          validate: validatePort,
         });
       }
     }
